+++ ffm/src/main/java/it/fantacalcio/ffm/domain/entity/Fido.java
@@ -24,7 +24,7 @@ public class Fido {
     private Squadra idSquadra;
 
     @Column(name = "IMPORTO", nullable = false)
-    private Byte importo;
+    private Integer importo;
 
     @Column(name = "TIPO_FIDO", nullable = false, length = 1)
     private String tipoFido;
